pyproject.toml
[project]
name = "kirppu"
description = "Second-hand sales POS and vendor signup for conventions"
version = "0"
requires-python = ">=3.10"

dependencies = [
    "django-environ~=0.11.0",
    "django-ipware~=5.0.0",
    "django-ratelimit~=4.0",
    "django~=4.2.0",
    "pillow~=10.0",
    "pubcode~=1.1.0",
    "mistune~=2.0.0",
    "whitenoise~=6.5.0",
]


[project.optional-dependencies]
oauth = [
    "requests~=2.30",
    "requests-oauthlib~=1.3.0",
    "oauthlib~=3.2.0",
]
production = [
    "gunicorn~=20.1.0",
    "psycopg~=3.1.0",
]
dev = [
    "factory-boy~=3.3.0",
    "Faker~=19.0",
    "pytest~=7.4.0",
    "pytest-django~=4.5.0",
    "pytest-env~=1.0.0",
    "pytest-cov~=4.0",
    "pytest-playwright",
    "pip-tools~=7.3",
]


[tool.isort]
profile = "black"
atomic = true
indent = 4
line_length = 120
no_sections = false
skip_glob = [
    "*/migrations/*",
    "/node_modules*/*",
    "/private/*",
    "/KirppuVenv*/*",
    "/venv*/*",
]
wrap_length = 120


[tool.pytest.ini_options]
addopts = "--tb=short -m \"not web\""
norecursedirs = "*.egg .* .git build dist node_modules* venv KirppuVenv* {arch} private kompassi*"
markers = "web: marks WebDriver/Selenium tests (deselect with '-m \"not web\"')"
DJANGO_SETTINGS_MODULE = "kirppu_project.settings"


[build-system]
requires = [
    "setuptools>=61.0.0",
]
build-backend = "setuptools.build_meta"


[tool.setuptools.packages.find]
where = ["."]
include = [
    "kirppu.*",
    "kirppu_project",
    "kirppuauth.*",
    "kompassi_crowd",
    "kompassi_oauth2",
]
